+++ VDWebPortal/Controllers/NewsMastersController.cs
@@ -8,6 +8,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using VDWebPortal.Models;
+using VDWebPortal.App_Code;
 
 namespace VDWebPortal.Controllers
 {
@@ -18,31 +19,52 @@ namespace VDWebPortal.Controllers
         // GET: NewsMasters
         public async Task<ActionResult> Index()
         {
-            var t_NewsMasters = db.T_NewsMasters.Include(t => t.M_SubCategoryMaster).Include(t => t.M_Master);
-            return View(await t_NewsMasters.ToListAsync());
+            if (!CommonFunctionVD.CheckUserAuthentication())
+            {
+                return RedirectToAction("Login", "Home");                
+            }
+            else
+            {
+                var t_NewsMasters = db.T_NewsMasters.Include(t => t.M_SubCategoryMaster).Include(t => t.M_Master);
+                return View(await t_NewsMasters.ToListAsync());
+            }
         }
 
         // GET: NewsMasters/Details/5
         public async Task<ActionResult> Details(int? id)
         {
-            if (id == null)
+            if (!CommonFunctionVD.CheckUserAuthentication())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Home");
             }
-            T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
-            if (t_NewsMasters == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
+                if (t_NewsMasters == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(t_NewsMasters);
             }
-            return View(t_NewsMasters);
         }
 
         // GET: NewsMasters/Create
         public ActionResult Create()
         {
-            ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName");
-            ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue");
-            return View();
+            if (!CommonFunctionVD.CheckUserAuthentication())
+            {
+                return RedirectToAction("Login", "Home");
+            }
+            else
+            {
+                ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName");
+                ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue");
+                return View();
+            }
         }
 
         // POST: NewsMasters/Create
@@ -50,35 +72,54 @@ namespace VDWebPortal.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "NewsID,NewsSubCategoryID,Heading,SubHeading,BodyMessage,VideoLink,Remarks,Location,NoOfViews,CreatedBy,CreatedDate,ModifiedBy,ModifiedDate,Active,NewsType")] T_NewsMasters t_NewsMasters)
+        public async Task<ActionResult> Create([Bind(Include = "NewsSubCategoryID,Heading,SubHeading,BodyMessage,VideoLink,Remarks,Location,NoOfViews,NewsType")] T_NewsMasters t_NewsMasters)
         {
-            if (ModelState.IsValid)
+            if (!CommonFunctionVD.CheckUserAuthentication())
             {
-                db.T_NewsMasters.Add(t_NewsMasters);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return RedirectToAction("Login", "Home");
             }
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    t_NewsMasters.CreatedBy = Session["EmailID"].ToString();
+                    t_NewsMasters.CreatedDate = DateTime.Now;
+                    t_NewsMasters.ModifiedBy = Session["EmailID"].ToString();
+                    t_NewsMasters.ModifiedDate = DateTime.Now;
+                    t_NewsMasters.Active = true;
+                    db.T_NewsMasters.Add(t_NewsMasters);
+                    await db.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
 
-            ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
-            ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
-            return View(t_NewsMasters);
+                ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
+                ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
+                return View(t_NewsMasters);
+            }
         }
 
         // GET: NewsMasters/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (!CommonFunctionVD.CheckUserAuthentication())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Home");
             }
-            T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
-            if (t_NewsMasters == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
+                if (t_NewsMasters == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
+                ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
+                return View(t_NewsMasters);
             }
-            ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
-            ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
-            return View(t_NewsMasters);
         }
 
         // POST: NewsMasters/Edit/5
@@ -86,32 +127,48 @@ namespace VDWebPortal.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include = "NewsID,NewsSubCategoryID,Heading,SubHeading,BodyMessage,VideoLink,Remarks,Location,NoOfViews,CreatedBy,CreatedDate,ModifiedBy,ModifiedDate,Active,NewsType")] T_NewsMasters t_NewsMasters)
+        public async Task<ActionResult> Edit([Bind(Include = "NewsID,NewsSubCategoryID,Heading,SubHeading,BodyMessage,VideoLink,Remarks,Location,NoOfViews,CreatedBy,CreatedDate,Active,NewsType")] T_NewsMasters t_NewsMasters)
         {
-            if (ModelState.IsValid)
+            if (!CommonFunctionVD.CheckUserAuthentication())
             {
-                db.Entry(t_NewsMasters).State = EntityState.Modified;
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return RedirectToAction("Login", "Home");
             }
-            ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
-            ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
-            return View(t_NewsMasters);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    t_NewsMasters.ModifiedBy = Session["EmailID"].ToString();
+                    t_NewsMasters.ModifiedDate = DateTime.Now;
+                    db.Entry(t_NewsMasters).State = EntityState.Modified;
+                    await db.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.NewsSubCategoryID = new SelectList(db.M_SubCategoryMaster, "SubCategoryID", "SubCategoryName", t_NewsMasters.NewsSubCategoryID);
+                ViewBag.NewsType = new SelectList(db.M_Master, "MasterID", "MasterValue", t_NewsMasters.NewsType);
+                return View(t_NewsMasters);
+            }
         }
 
         // GET: NewsMasters/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (!CommonFunctionVD.CheckUserAuthentication())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Home");
             }
-            T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
-            if (t_NewsMasters == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
+                if (t_NewsMasters == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(t_NewsMasters);
             }
-            return View(t_NewsMasters);
         }
 
         // POST: NewsMasters/Delete/5
@@ -119,10 +176,17 @@ namespace VDWebPortal.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
-            db.T_NewsMasters.Remove(t_NewsMasters);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            if (!CommonFunctionVD.CheckUserAuthentication())
+            {
+                return RedirectToAction("Login", "Home");
+            }
+            else
+            {
+                T_NewsMasters t_NewsMasters = await db.T_NewsMasters.FindAsync(id);
+                db.T_NewsMasters.Remove(t_NewsMasters);
+                await db.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
         }
 
         protected override void Dispose(bool disposing)
